% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory.Storage.Service.R
\name{Memory.Storage.Service}
\alias{Memory.Storage.Service}
\title{Memory Storage Provider}
\value{
a list value:
\itemize{
\item \code{storage <- broker |> Memory.Storage.Service()}
\item \code{storage[['Seed']](data, table)}
\item \code{storage[['ExecuteQuery']](...)}
\item \code{storage[['Insert']](entity, table)}
\item \code{storage[['Select']](table, fields)}
\item \code{storage[['SelectWhereId']](id, table, fields)}
\item \code{storage[['Update']](entity, table)}
\item \code{storage[['Delete']](id, table)}
}
}
\description{
Provide a common interface to perform standard data operations on an in-memory data store
}
